Introduce a "repoUserValidationMethod" client setting
Used in UpdateRepo to ensure the user on client is the same
as the one on repo with the same name.

Supports two modes:
* centralauth: Make use of CentralAuthUser.
* assumeSame: Assume that if a user with that name exists on
    repo, it will be the same.

This way we might be able to remove the binding against CentralAuth,
now that the SULF has been finished.

Also needed in order to be able to write tests for
UpdateRepoHookHandlers without having to work with CentralAuth.

Change-Id: I2f9c0c5743a66b1e62f633ee3c64dad407fef424