Update patch set 1
Patch Set 1:

I had this situation: the WikibaseClient class is known to the class loader, but the default configuration is not loaded. Some code (in my example this is code from ArticlePlaceholder) calls WikibaseClient::getDefaultInstance. It starts doing stuff but fails at a totally arbitrary place with an error that does not tell anything about what happened and what could be done to resolve it.

These two simple checks guard against multiple kinds of misconfiguration. They don't hurt - they are called exactly 1 time.

An if() checking a constant at this point feels extremely weird, if not wrong. No code in WikibaseClient depends on this constant, or does it?

Patch-set: 1