Add missing tests for setOptions methods in all validators
These methods are currently dominating the top 10 "Least Tested Methods"
in the PHPUnit coverage report. Sure, these tests don't give us much.
But it also does not hurt to have them. The most important effect is:
With these tests the coverage of about 15 classes jumps to 100%. For
free. This will raise the class coverage in repo from currently 29% to
about 35%.

Change-Id: I24fecf1dcc4a86bd33655bb85e38dac3a1a82e5f