Update patch set 3
Patch Set 3:

As a stop gap this should be ok, but we need to make sure we don't forget about the description changes. Ideally, another patch with a better solution would already be in the pipeline when we merge this.

Or at least put big FIXMEs not just into ChangeFactory, but also into AffectedPagesFinder, to avoid confusion about receiving incomplete diffs.

Patch-set: 3