Add public item/property (de)serializer factory methods
These methods need to be accessible from outside because we want to create
a Dispatchable(De)Serializer in a place where we know all available entity
types and have access to their serializers. That way we can avoid
unneccessarily nesting dispatching serializers and have a clear place in
our code where that dispatching happends.

Bug: T126932