Update patch set 1
Patch Set 1:

We also call EntityDocument::isEmpty, which was introduced in 4.3. But this is not critical, because all implementations always supported this method. Nothing will explode when new code is run with DataModel 4.2. I suggest to stick to that for now and update to 5.0 later.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0