Update patch set 1
Patch Set 1:

A -2? Seriously? What's so bad about this that a merge must be prevented?

Empty terms are not ignored in the DataModel version the current Wikibase.git master uses. I'm sure I explained that in the commit message.

I5c81219 was a breaking change. Not horrible. The worst thing that can happen is some wasted bytes in the serialization.

Remember when you said nobody urgently needs DataModel 5? This is actually the reason we wanted it fast.

Patch-set: 1