Update patch set 5
Patch Set 5:

Huh? Did I suggested this? You don't need to check for Fingerprint in EntityContent if you don't want to. And even if - sure, that would mean code sharing, which should be avoided, but it is also not good practice to avoid it at all costs.

A possible solution is to simply let subclasses override getTextForSearchIndex and getTextForSummary instead of basically duplicating these methods with slightly modified names. Is the benefit these additional methods give us and other extensions implementing new entity types really worth it? I tend to say no, it isn't.

Patch-set: 5