Allow injection of entity views
This patch renames EntityViewFactory to ViewFactory and introduces a new
EntityViewFactory. This EntityViewFactory contains map of entity types to
callbacks that create a new EntityView given a ViewFactory and some
additional parameters. The callbacks are currently defined in WikibaseRepo
but can be moved easily to a central definition once it gets merged.

Bug: T127191
Change-Id: If7e3d73dbc89c83f60d5b6acc53e1819eb4db630