Allow injection of entity views
This patch introduces ViewFactory and BasicViewFactory. While
BasicViewFactory provides all services that are the same for every
request, ViewFactory is a wrapper of it and provides all other
dependencies that are only available at runtime.

Factory callbacks are passed to EntityParserOutputGenerator that allow the
instantiation of specialized EntityView implementations. They are
currently defined in WikibaseRepo but can be moved easily into the central
registry once that gets merged.

Bug: T127191
Change-Id: If7e3d73dbc89c83f60d5b6acc53e1819eb4db630