Update patch set 7
Patch Set 7:

At least not in this repository. When all patches on the current sprint are merged, Entity is not used in production code of Wikibase.git anymore (tests still need updating but are no problem as long as we still have Entity in the DataModel).

Patch-set: 7
Reviewer: Bene <549@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0