Fix bug in EntityParserOutputDataUpdater
This fixes an actual bug in the data updater class which doesn't register
SiteLinkDataUpdater instances if they are also an instance of
StatementDataUpdater.

By fixing this, we can get rid of the special handling for badges in
EntityParserOutputGenerator as it only duplicates the work done in the
EntityParserOutputDataUpdater and ReferencedEntitiesDataUpdater.

Change-Id: I869e73324eca73b2cc9396bf27db0774feb02c48