Update patch set 1
Patch Set 1:

What works for me is the following piece of code:

.wikibase-sitelinkview-page {
    white-space: nowrap;
    overflow: hidden;
    text-overflow: ellipsis;
    max-width: 300px;
    display: table-cell;
}

It displays nicely for up to 3 badges and I don't think there are sitelinks with more than 3 badges. With more badges, it does a line break but we perhaps have to adjust the interface anyways if we are going to have more than 3 badges on sitelinks. It is a bit ugly though that when no badges are set there is still some whitespace unused.

Patch-set: 1