Update patch set 2
Patch Set 2: Code-Review-1

This test needs to be fixed to not duplicate and test the fall back definition in core. So we either need to mock the calls obtaining the fall back definition. Or load the fall back definition from core instead of duplicating it for the expected value in the test.

If we don't do this then these can neither be changed in core nor Wikibase without breaking the other.

Patch-set: 2
Reviewer: JanZerebecki <640@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1