Set @codeCoverageIgnore on ApiBase::getExamplesMessages methods
How to write a test for this protected method? Sure, we can come up with
a setup, but it would be so complicated and slow, I don't think that's
worth it. What would it test? Probably that it's an array, not empty, and
only contains string keys and values. But these formal details aren't
worth much. This is about content, but content can not be validated easily.

Change-Id: Ia55b440fdd3c0f3f7262ca6d576b983849f33f11