Update patch set 2
Patch Set 2:

I added a test for the "sitelinks not supported" case. It is not possible to test the other cases at the moment because all entity types supported by Wikibase currently have a fingerprint and statements.

We could insert a mock/fixture but that would require quite some refactoring of the api class and the test while we also need to somehow inject fixture serializers etc. That seems to be doable but goes beyond this patch and requires some work.

Patch-set: 2
Reviewer: Bene <549@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0