Update patch set 3
Patch Set 3: Code-Review+2

I agree that the error code change is a bit unmotivated, and technically a breaking change. I don't think any clients are relying on the code though, since the error case that was covered by the old code was easily anticipated on the client (adding sitelinks to properties). Basically, if the code was smart enough to check the code, it should be smart enough to avoid the error.

Patch-set: 3
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2