Implement usage tracking without eu_touched
Per T124737.

This leaves a few fixmes in EntityUsageTable which I'll
happily address in a follow-up once this has received
some code review.

While we stop using or explicitly setting the eu_touched
field it is not yet removed from the database schema.
This should be done in another change.

I've manually tested this and also manually tested upgrading
and downgrading. The changes to AddUsagesForPageJob are
fine with upgrading (jobs created with the old version
can be run by the new version), but jobs created with the
new version will break in the old version. Given that is
rather low impact and will only happen in case we revert
the branch, this should be acceptable.

Because the timestamps weren't really used anywhere (see
the ticket) the tracking data written by the new version
works fine with the old one (even though eu_touched is
empty). I manually tested that.

Bug: T125838
Change-Id: Ib9325237691e853a2e4c4f7d9d202c502653395b