Rename client folders to be PSR-4 compilant
I want to do this in multiple small steps: This patch does not touch
the namespace headers of these files. All it does is renaming folders.
That's all. This can not break anything (because of the way these
classes are currently loaded) and will make later PSR-4 related patches
much easier to review.

"api" is renamed in an other patch, see I7fd2033.

Change-Id: I1dda9de090f80e06dc57ec3bb42d294a2bc551d1