Remove ChangeOpBase::validate from abstract base class
The base class seems pointless. Why should there be a default implementation
for a validation that's always successfull? This is called from a few
subclasses, but not used in the sense that a subclass defaults to this
implementation. They all have their own validate method.

The remaining updateSummary method is a helper function. Code sharing. But
I did not wanted to touch this in the same patch. Will continue working on
this in an other patch.

Change-Id: I7835c0d6150c93ae56181c7bf341d80a8013a353