Add strong array type hints to function parameters
I found these with a regular expression:
(?s)@param\s+(array|\w+\[\])\s+(\$\w+\b)(?![^{}]*\barray &?\2\b)

This means: When a function parameter is documented as "@param array"
or "@param string[]" or something, but there is no hard "array" type hint
in the function header, I checked and added the hint when it looks like
the documentation is correct and this can't be anything but an array.

I was a bit suprised that there is not more of this left in our code base.

If you find something suspicious in this patch, please tell me and I will
split it of to a separate patch.

Change-Id: If12f0265c41c4648bee639dd788d18ef00950c29