Replace non-tests for setOptions with codeCoverageIgnore
I was under the impression I fixed this mistake I made already. Today
I saw this as a huge list of "risky tests" in a Jenkins log. This was
not my intention to make this *more* visible. I wanted to remove these
uncovered methods from the code coverage report. There is nothing to
test here.

Change-Id: I6280fda32da8ba6419790e9a7643d8db73ea9f83