Update patch set 1
Patch Set 1:

What's going on here? How is it possible that this error was reported 4 (!) times before, and with this fix it's still reported 2 times? With old (!) code that still fails for exactly the reason I'm fixing in this patch. How is this possible? This is only 1 test with no provider. It should be executed 1 time only. And never with two different versions of the same file.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0