Update patch set 10
Patch Set 10:

I'm not doing changes in a rush. There was no need to. assertTag is still there and can be used. My patches had the intention to slowly migrate away from it, and ideally making tests and assertions easier to read. So yes, all intended. Please do not revert.

Patch-set: 10