Don't call setters when not necessary
This replaces lots of usages of ...Holder interfaces with their respective
...Provider interfaces and doesn't call the setters anymore. While the
provider interfaces don't give a guarantee that changes to the returned
object will be reflected in the provider itself, all known implementations
supported by Wikibase meet this condition. Furthermore, DataModel 6.0 will
make this requirement clear in the interface declaration.

Bug: T128362
Change-Id: I21ccea0c65111c33f105b16429de0a6b89e295bf