Update patch set 6
Patch Set 6:

I tried to fix this test for an hour or so and gave up. What I understood it does: it adds a new sitelinkview to the sitelinklistview and then expects events on the sitelinkgroupview. I do not think this behavior should be tested, at least not like this.

Patch-set: 6
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0