Shorten out of JavaScript widgets value() methods when possible
Note that this shorten-out is *not* possible for all widgets. This
depends a lot on what the value() method does. In the cases I'm touching
here the value() method checks for isInEditMode and then returns
this.options.value. This means: in this case the value is compared with
itself. Pointless. Shorten-out.

Change-Id: Ic5eeec3101bc974eb39071cd20a17b145c08468d