Update patch set 6
Patch Set 6:

(1 comment)

This needs a corresponding change in https://github.com/wmde/WikidataBrowserTests/ and it should keep the environments and env files in sync. Otherwise it is not possible to move scenarios from WikidataBrowserTests to Wikibase. (Yes I didn't do that myself when last changing them. I think I kept the result of environments compatible, so the sync can easily happen.)

Patch-set: 6