Update patch set 1
Patch Set 1:

This seems ok as it is, but it would make sense to put the definition of the ContentHandler factory callbacks into the definition array at the same time (or into an immediate follow-up).

Baslically, there should be a 'content-handler-factory-callback' key that would contain the callbacks we currently put into $wgContentHandlers directly in Wikibase.php. 

We can then copy the callbacks from the EntityTypeDefinitions object to $wgContentHandlers. I'm not sure when to best do that, though. Perhaps it would be better to implement the ContentHandlerForModelID hook.

Patch-set: 1
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0