Update patch set 8
Patch Set 8:

I think it's fine to have extra settings for CI. But these are really specific to the CI environment, so they shouldn't live in the extension repo. 

This change is mostly a quick fix. The alternative is to move the top level Wikibase.php and the settings files it uses to a separate repo (or to the repo that has the jenkins config)

Patch-set: 8