Update patch set 8
Patch Set 8: Code-Review-1

Having special settings for one repo for CI in a different repo is even worse. We wouldn't have the tests in a different repo, right?

Why would you want to test in a different configuration than you normally run? It is not like we try to rig an emissions test.

Patch-set: 8
Label: Code-Review=-1