Update patch set 8
Patch Set 8:

@JanZerebecki Thinking about this, what do you mean when saying we should test the code with the same settings that it "usually" runs? We certainly can't use the config that is used on the live wikimedia cluster, especially not the database setup, or the live job queue, or the live caches.

So, we want to use a config similar to a typical 3rd party setup. But with some differences in the configuration of the database access, caching, and generally anything related to external services and resources. These things need to be overwritten for the testing environment.

Patch-set: 8
Label: Verified=0