Update patch set 10
Patch Set 10:

Even external services like a database, etc will get configured automatically in the grand future. But I agree it shouldn't be the same as the live database, that in development environment you use the same settings file for the phpunit run as the normal installation is a bad idea. It should expect LocalSettingsTests.php or something. Things that are not services that the tests can set up itself they should do and not expect a human to do that manually.

Patch-set: 10
Label: Verified=0