Update patch set 10
Patch Set 10:

Btw. phpunit.php sets up many things the tests need, which is better than in the code that will run in production. The part that is bad about that is that it always does that and with a global context. Ideally it would do so only when needed only available to the scope of that one test.

Patch-set: 10