Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

This confused me. It looks like this only moves existing code around, but the inverted conditions scare me.

In the old code, draw() initialized the notification, set the content to nothing and appended it to the header.

In the new code, draw() does not do this any more. Instead, the notification() function takes care of this initialization.

Additionally, two more lines became initialization code (was update code before). If they were misplaced, I wonder how this happened.

The possible issue is that the notification() is now reseted every time draw() is called.

Patch-set: 1
Label: Code-Review=-1