Update patch set 11
Patch Set 11:

I think for the cases where it is ok to have a setting specially for running under test this patches files can not know the setting, because it is set up externally. E.g. the db connect string.

For something like travis where the CI is described in the same repo in travis.yml that file knows the db connect string and needs to inject it, but nothing else. And that file is neither part of the code under test like what we modify here nor part of the test code. The test code needs it to be injected.

Patch-set: 11
Reviewer: JanZerebecki <640@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0