Use PHP 5.5's …::class feature instead of strings in client
I2c1a4d6 does the same in view. More patches to come (there are ~90
full qualified class names in strings in lib, and ~280 in repo).

This is not a blind, fully automatic replacement. I reviewed every
replacement. I found a mistake while doing this (see
DataUpdateHookHandlersTest) and replaced SiteLinkTable mocks with
the SiteLinkLookup interface when possible.

Change-Id: I577208e75657864d2badb636659f87908fedf794