Add cache purging for flavored Special:EntityData output.
This causes URLS with all possible combinations of parameters
to be send to the caching layer for purging. As of this patch,
this means we have to purge four flavors for six formats, so we
purge 24 urls instead of the 6 we purged before.
Whether this is necessary depends on if and how the caching layer
supports "buckets", and how they are configured.

Bug: T128667
Change-Id: I6c6d093f5af164e3a4df4493f3aed15d82bb9741