[WIP] Send a notification when a page is connected to a repo item
When a page is connected to the associated repository by a user different
from the one who created the page, an Echo notification is sent to the user
on the client wiki. One of the secondary links is customisable similarly to
how welcome notifications are. The web notification is bundled.

This is configurable per client and requires Echo extension enabled.

The remaining point to be discussed is what icon should be associated with
the notifications.

Bug: T110604
Change-Id: I2a3babd7f5bcc0cbf923d94b1ea3fe768457a95d