Update patch set 3
Patch Set 3: Code-Review-2

(3 comments)

I'm still not entirely convinced we want or need this. If I understand correctly, the idea is to have a "preferred content language" separate from the UI language. The preferred content language would be defined solely based on Babel boxes, right? And used to decide which terms to show in the term box? How would I then switch to different terms being show?

I'd like to see a product level discussion and decision about this. So far, the preferred content language was always derived from the current UI language. Now we have to find out what to do when this isn't possible, because the UI language isn't a valid content language.

CR-2 because this is blocked on a product decision. The code seems fine on a cursory look.

Patch-set: 3
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2