Update patch set 6
Patch Set 6:

The problem with PURGEs is that they are mostly invisible to developers. We have had code changes that have caused massive fluctuation in PURGE volumes, and more often than not we are not able to pinpoint the source of the change. How often are these URLs hit, in practice? Would it be viable to just make them uncachable for now?

Patch-set: 6
Reviewer: Ori.livneh <255@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0