Make sitelinklistview.tests.js more specific
This is a direct follow up to Ie665dac. I found the messages not
sufficent. The key is: when you add a new item to the list, it's
critical if the new item is edited or still in it's initial state.

The bugfix applied in Ie665dac adds an !== check. This means, it
does not care about the state of the individual items when the
total count is different. This is fine and what the test now
explicitely tests.

Bug: T129450
Change-Id: I37cf69709f8d9874c5a34d7d7e7aa3cef4c09e61