Update patch set 5
Patch Set 5: Verified-1 Code-Review+1

Looks good so far, but needs an update of the "@dev" dependencies in composer.json.

Question: What is the reason we have to drop compatibility with the previous DataModel version? My impression is that this patch here is still compatible, but the next one makes use of the new interfaces in Item and Property. This will be incompatible. Is this correct?

Patch-set: 5
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=-1
Label: Code-Review=+1