Update patch set 1
Patch Set 1: Code-Review-1

This seems extremely specific to an extremely specific use case that should never happen. Why does this need to be a template? This will neither be reused in other places of the code nor reused in the frontend, as far as I can see. I would not clutter the templates with this but simply inline it as a static string. This is not worse as it is now. This Html::… stuff is barely more than a string builder anyway.

Patch-set: 1
Label: Code-Review=-1