Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

The change is supercool. I would +2 this, but I would like to continue refactoring this a little bit, as explained in the comments. Please tell me what you prefer: Is it ok for you when I do this in a later patch; or should I amend this patch; do you want to do this; or something else?

Patch-set: 2