[WIP] Introduce EditEntityHandlers for wbeditentity
This patch removes the last and only real purpose of EntityFactory which
is in the wbeditentity api module (EditEntity). The interface currently
contains two methods to create new entities based on a data array or an
existing entity (as a replacement for clear).

We might want to factor out even more of the code that lives within
EditEntity at the moment and add more methods to the interface but for now
this is sufficient to allow some basic new entity types like mediainfo.

Change-Id: Icd0ad807aca6b4fe037f38631369e662b243e5af