Update patch set 1
Patch Set 1: Code-Review-2

(5 comments)

This change only shifts the type-unsafety from the code that constructs the SnakSerializationRenderer to the code inside SnakSerializationRenderer. We actually lose an early check with a useful stack trace if we do this.

The issue is caused by the overly generic Serializer/Deserializer interfaces, which are not really useful without generics or some additional mechanism for making sure we have the right kind of serializer. I suggest to avoid using these overly broad interfaces, and instead typehint against more narrow interfaces wherever possible. This means doing a "narrowing cast" as far up the stack as possible (in the case of PHP, a narrowing cast is just /* @var Foo $foo */)

Patch-set: 1