Remove prev/next rotation buttons from listrotator
This intentionally is a small step, so each patch is as small as possible
and always leaves the UI in a non-broken and non-broken looking state.

This removes the two extra UI elements left and right from the dropdown
menu. With this change this is not really a "rotator" any more, because
you can not rotate anything. It's a simple dropdown. But as said, I want
patches that are as small as possible, therefore no name change.

Bug: T109455
Bug: T109457
Change-Id: I6a37f34e73f476da4b5e164071deb99515468ebf