Update patch set 1
Patch Set 1: Code-Review+1

Per Jeroen, using the classes as done now violates the contract of the data-model-serialization package. For example, changing the name of a package private class isn't a breaking change but would break the usage in this component. The current code in Wikibase.git is wrong and no matter how we decide later, we should merge this patch to fix the violation of the contract of the current version of data-model-serialization.

Patch-set: 1
Reviewer: Bene <549@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1