Don't assume English in MwTimeIsoParser
In order for this to work we need to expand {{PLURAL:$1|…}}
constructs which many localizations make use of.

Getting this to work for the tests wasn't nice, but I figured
that getting the message directly from the Message object would
work.

This most probably still has some shortcomings, but I think it
works reasonably well (like the English version).

Bug: T95532
Change-Id: I15c2864ce9a2b6ef971982da5963347d60835992