Update patch set 1
Patch Set 1:

Personally I am not 100% against this change then, but it may make more sense to simply add this method to the class in WikibaseMediaInfo too.
This may mean less people are tripped up by the cross extension dependency (even thought it is just in tests).
On the flip side I don't mind that much but thought I should point out that option.
If nothing happens here in the next week give me a poke!

Patch-set: 1