Update patch set 1
Patch Set 1:

I'm fixing an edge case that currently does not have an effect, because empty aliases are ignored anyway when constructing the final AliasGroup object. But I'm doing it anyway to keep our code base as clean as possible. We call this refactoring.

In other words: Your code worked, but I felt it was a bit fragile, so I tried to make it more robust. This does not mean there was anything wrong with what you did.

Patch-set: 1