Assume users with the same name on the repo and client are the same
This way we are be able to remove the binding against CentralAuth,
now that the SULF has been finished. The collateral damage caused
by this should be very low to non-existant, given that virtually
no one except for the Wikimedia foundation uses CentralAuth…
and those that do probably don't have Wikibase installed.

Also needed in order to be able to write tests for
UpdateRepoHookHandlers without having to work with CentralAuth.

In case this behaviour is not desirable, UpdateRepo can be disabled
completely by setting the "propagateChangesToRepo" client setting
to false.

Change-Id: I2f9c0c5743a66b1e62f633ee3c64dad407fef424