Remove animation from statementgrouplabelscroll.js
My first attempt was to make the animation faster. But while playing
around with this I realized this code is just broken and bogus. The
animation is canceled each time a "scroll" or "resize" event fires.
When you scroll very slowly, the animation never finishes.

I suggest to remove it. The property label jumps to the new position.
Done.

Change-Id: I1962d96a7a087a5fb8e90bc49bc84fff4855dc68