Replace statementgrouplabelscroll animation with CSS transition
The JavaScript driven animation was slow and bogus. The animation
was canceled each time a "scroll" or "resize" event fired. When you
scrolled very slowly, the animation never finished.

CSS is so much reliable.

In very old browsers the property label jumps to the new position
with no animation steps inbetween.

Bug: T103485
Change-Id: I1962d96a7a087a5fb8e90bc49bc84fff4855dc68