Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

I'm still afraid of dangerous edge cases in the code that builds the regex. I suggest to do something like this:

Match \{\{PLURAL\:\$1(\|[^{|}]*)+\}\} with a regex. Then do all the replacements in this substring only. This guarantees two things: No closing }} that was not part of a PLURAL function is turned into dangling brackets. And no pipe that is not part of a PLURAL function is unescaped.

Patch-set: 3
Label: Code-Review=-1