Update patch set 2
Patch Set 2: Code-Review-1

This patch doesn't actually rename the class, as it should. Perhaps you forgot to `git add` something?

I agree that LocalMirrorStrore the the better name. 

However, the class in question is disused and probably broken by now. 

It is intended to support clients (and federated repos) that cannot be accessed via a direct database connection. It would certainly be nice to support that, but we'll have to do quite a bit of refactoring to make that work. My gut feeling is that at the moment, it would make more sense to just remove CachingSqlStore instead of renaming it.

Patch-set: 2
Label: Code-Review=-1