Update patch set 3
Patch Set 3: Code-Review+1

There is no class with this name in our code base. The only 2 classes implementing the ClientStore interface are DirectSqlStore and MockClientStore. I think this was planned and never implemented, which makes this more a TODO that should probably not be in this documentation.

Patch-set: 3
Label: Code-Review=+1