[WIP] Add @return …|PHPUnit_Framework_MockObject_MockObject tags
The pipe usually means "or", but is used as an "and" here. This is a
code smell. It supresses type safety warnings in tools like PHPStorm
but does not really make the code better, because now the methods are
allowed to return *any* mock. Solution: Refactor the code so callers
do not need to know the methods return mocks.

Change-Id: I3277f220c0082878c237e809e0220edfeab57f4c