Replace empty array()s with PHP 5.4 [] syntax in purtle and view
This specific details bugs me since I started coding PHP 3: That you
can not declare arrays with $var = [].

I suggest we do this in bulk, as a first step (only touching empty
arrays) towards the short array syntax. Please add a +1 if you are fine
with this. I will do the same in repo, client and lib then.

Change-Id: If5a2477db4755ffa62acd49273c6c469d0034b3e