Remove SiteList from GetEntities API for preformance
The GetEntities is called millions of times a day. This means this code,
fetching about 1000 sites and sorting them alphabetically for convenience
in the ApiSandbox (see SiteLinkTargetProvider::getSiteList), is called
millions of times, but not needed in aprox. 99% of the cases.

WARNING: This is a breaking change because the API does not report invalid
site ids any more. Instead you just get an empty result. Personally I think
this is not a problem.

Change-Id: I952e6c29b418c307c0974dfd75a1267342df1018