Update patch set 3
Patch Set 3:

I think having an (int) inside the concatenation is good, because that way it's obvious that under no circumstances escaping would be an issue there.

Having an integration test for this would be nice, but I don't know how you would go about writing it... how shall we assert that Maria uses the right index? For that we would need to capture the query and EXPLAIN it… don't think that's worth it.

Patch-set: 3
Reviewer: Hoo man <50@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0