Update patch set 8
Patch Set 8: Code-Review-1

(1 comment)

We (Jonas and I) did a code review. We *think* this patch can be merged as it is, but we have lots of comments and suggestions and uploaded a separate patch (on top of this) that renames and refines all new classes introduced in this patch. Please have a look and decide: You can change your patch and move it towards our suggestion, or you could +2 our patch and we merge both as they are.

Patch-set: 8