Update patch set 1
Patch Set 1:

There is no default in Dummy.

It's not correct to state that a dummy string like "(brackets)" is English. The correct language code for this is qqx.

I don#t see how you can "disagree" with the fact that message parameters can be integers. This is already the case. All I do is documenting the fact.

Patch-set: 1
Label: Verified=0