Update patch set 1
Patch Set 1:

And as for parameter types: This is a new interface. WE are designing it. We can decide what it accepts. It may be more restrictive than one or even all implementations technically are. That doesn't change the fact that this interface is designed to only take strings and currently only takes strings in production code. The only support for non-string parameters we will have is silent conversion to strings.

Patch-set: 1