Update patch set 2
Patch Set 2:

It's not forbidden to use qqx here. This is in the "reserved for local use" range in ISO 639 and we are allowed to use a code that "accidentally" happens to be the same as in core.

I'm fine with zxx if this turns out to be a blocker.

Patch-set: 2
Label: Verified=0