Add missing title tooltip to snak type selector
This feels a bit pointless when the snak type is set to "no value"
or "some value", because there is already a big "no value" or
"some value" clearly visible in the UI. But it's not obvious that
this text describes the icon. When set to "custom value" the meaning
of the icon is not visible at all, until you click it.

For consistency and the reason described above I want the tooltip in
all 3 states.

Change-Id: I8793e63a25e583a93e5af9893cf8c368a63dfe6f