Rename folders in lib/tests/phpunit/… to upper case
This is an absolute minimal patch that does nothing but renaming folders
to start with an upper case letter. This will be needed for PSR-4
compatibility.

There are some lower case folders left. This is done on purpose. This
patch focuses on cases I find obvious. Other folders need more work,
e.g. files may need to move to other folders.

Change-Id: I20421343b5da927f7ce5aa761452400577d72b39