Bring lib/tests/phpunit folder back in sync with lib/includes
This is the most minimal patch to make the tests folder mirror the
structure of the code it tests. This could also be done the other way
around, by moving all test files to an "includes" sub-folder. I decided
to do it this way instead because I find the mirrored "includes" folder
pointless.

I'm doing this to make Wikibase.git's folder structure more compatible
with PSR-4.

Change-Id: Ie75e92be7691369f96e8c3a11d799b2d16ac6b26