Reduce EntityFactory implementation and usage to an absolute minimum
The interfaces this class currently provides are either not used at all
or stuck to a very specific usage that can not be expanded easily. We
know we must redo this. To make this way forward much easier I'm removing
all code that's either not used at all or not flexible enough to build
upon. All I keep is:
* The class itself. We probably need an EntityFactory.
* The newEmpty method. It's the only one that's currently used.

Bug: T132964
Change-Id: I6871b1979e54fca2f571e77baff6870323c5c73f