Update patch set 1
Patch Set 1:

I really can't tell what the impact of this version bump is.

Ricordisamoa, I need more information of what the intention of this patch is. It's much easier to +2 this when we know, for example, that we are the last extension using these old versions.

Palodox, what exactly does your comment mean? Should we wait for a bugfix in grunt?

Patch-set: 1