Update patch set 8
Patch Set 8: Code-Review-1

(3 comments)

It doesn't feel right that we have to do this. It should be possible to purge all URLs that start with the same prefix, no matter what parameters are added. I know this is not so simply with the infrastructure we are using. This is not the reason for my -1.

See comments for reasons for my -1.

Patch-set: 8
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1