Add XML header to all SVG files
ResourceLoader is not able to detect the correct content type when this
header is missing. It inlines the SVG with the wrong MIME type
"text/plain" instead of "image/svg+xml". The browsers are not able to
understand this (tested in Firefox).

I will check if this is a bug in ResourceLoader that can be fixed.

Change-Id: I62a029f1a69399fe5c2183055ef7a3f55e137e66