Update patch set 1
Patch Set 1:

(1 comment)

The -1 is for "this should use information from EntityTypeDefinitions". If that happens in a follow-up, fine - but then I'd like to have a look at that follow-up first.

That's the downside of small patches: it's hard to see "why" something is done, or is done in a particular way.

Anyway, it seems to me that it would be easier to change the existing code to use callbacks, instead of removing the registry-style code, and then later re-adding it.

Patch-set: 1