Update patch set 1
Patch Set 1:

Sorry, but no. The "why" is as obvious as it can be. All this patch does is removing unused and redundant code.

You are, again, misusing the -1. It literally means: "there's a problem with *this* change, please improve". But you even said it: There is nothing I can do on *this* patch to make you happy. I don't want to play this game. I'm not good at it. I kindly ask you to stop gaming this system.

Patch-set: 1