Update patch set 13
Patch Set 13: Code-Review-1

Seems like this puts all the Site objects into the ACCELL cache *again*. That's a LOT of data. I'd rather want to avoid that.

If we want to cache the sorting/grouping, just have a sorted & grouped datat structure with site IDs, not full site objects.

Or we just stop using the actual list of sites in the API pram type definition, and allow any string.

Patch-set: 13