Update patch set 1
Patch Set 1:

I made an alternative patch, I366da6d41 that goes directly from a class based registry to a callback based registry. I think it is easier to read and review, because the intent is immediately clear. Also, it's smaller by itself, and no second patch is needed to make it used callbacks. The next step is to make the callbacks configurable via EntitytypeDefinitions.

Patch-set: 1