Update patch set 15
Patch Set 15:

As both the presentation model and the formatter extend Echo classes which are not loaded during tests, testing them seems to be impossible. Maybe someone can think of a workaround... Is it possible to load Echo inside tests? If not, is it okay to supress tests when Echo is not loaded (but it can be locally)?

Patch-set: 15
Label: Code-Review=0