Update patch set 2
Patch Set 2:

This is not an "alternative". It's almost the opposite of what I wanted to achieve with the intentional split I did in I6871b19. This patch, as it is now, leaves the code in an utterly dirty state. Inconsistent; lots of unused code; some outdated and lots of non-informational, redundant comments. What I find most silly is that the hardcoded list of entity types is *still* hardcoded at the *exact* same place.

What this patch really does is *replacing* the factory with a completely different implementation. It's like a step sideways instead of what I6871b19 does: go a *small* step backwards (it's really a very, very small step because all the removed functionality is unused or hardcoded anyway), making room for a big step forward. With this patch here that proposed step forward is no fun because you have to maintain, update and keep all that unused and outdated stuff in sync. I really, really did not wanted to waste more of our time with this silly little thing. That's why I killed it, hoping for a two minutes review. Then there would have been lots and lots of room and fresh air for improvement.

With this patch in place we are digging dust.

Patch-set: 2